add AOF Reader and Writer
contains functions to read and write to an AOF file.
the reader takes bulk string commands from a file and
executes them.

the writer writes bulk string versions of sent commands
to the file ready to read on next startup.
this is owned by the server and created during init.

--

build.zig: ZLS check step for better errors
- see: https://zigtools.org/zls/guides/build-on-save/
main.zig: write error name and not bytes
.gitignore: aof file