decouple commands from Client
this removes the need for commands to have a Client
sent to them, instead only taking their minimum
requirements to function.

there are now three different ClientHandlers to accompany this:
Default, Client, and Store.

by only taking their minimum requirement, the commands
are able to be called from things that aren't a client
(such as AOF), and only need a writer as opposed to a
network stream.